album_folder option to move
Modified this album_folder.py to accept a --move option and require the source/dest dir. 
THE DEFAULT IS TO COPY, as the original script. 
This was done because the original is only to copy, and since my library has 800Gb and I don't have that extra space to copy to another HD, my only option is to move all files to the albuns instead of copy then. 

hope helps somebody. 
Not tested on my library yet, but the code should work.