Add files via upload
This new version  has a lot of changes. I realized that in the last version, just moving the file didn't work, because if the photo is included in several albuns, the other copies will not be done. So, this new version first copies the photo to all other albuns, and then it moves the file to the last album listed.